+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Updated `github.com/evanphx/json-patch` to v5.6.0.
 - Updated `github.com/go-logr/logr` to v1.2.4.
 - Updated `sigs.k8s.io/yaml` to v1.3.0.
+- Optionally allow annotation name different than LastAppliedConfigAnnotation [#97](https://github.com/manifestival/manifestival/issues/97)
 
 ### Added